% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HusbyFieldSheet.R
\name{HusbyFieldSheet}
\alias{HusbyFieldSheet}
\title{HusbyFieldSheet}
\usage{
HusbyFieldSheet(
  path,
  SheetNames,
  parallel = FALSE,
  cores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{path}{character, the file path of the excel sheet}

\item{SheetNames}{character vector, sheet names of the excel sheet}

\item{parallel}{logical, whether to run the function in parallel}

\item{cores}{integer, number of cores to use if parallel = TRUE}

\item{verbose}{logical, whether to write messages while the function runs}
}
\value{
dataframe, a dataframe containing the data from the excel sheet
}
\description{
This function reads excel files with sheet names and returns a data frame
}
